+++ src/main/java/com/umc/finly/domain/home/controller/HomeRecordController.java
@@ -1,6 +1,6 @@
 package com.umc.finly.domain.home.controller;
 
-import com.umc.finly.domain.home.dto.res.HomeRecordsResDTO;
+import com.umc.finly.domain.home.dto.response.HomeRecordsResDTO;
 import com.umc.finly.domain.home.service.HomeRecordService;
 import com.umc.finly.global.apiPayload.response.ApiResponse;
 import com.umc.finly.global.apiPayload.response.SuccessCode;
