+++ src/main/java/com/umc/finly/domain/home/converter/HomeRecordConverter.java
@@ -1,6 +1,6 @@
 package com.umc.finly.domain.home.converter;
 
-import com.umc.finly.domain.home.dto.res.HomeRecordsResDTO;
+import com.umc.finly.domain.home.dto.response.HomeRecordsResDTO;
 import com.umc.finly.domain.record.dto.res.RecordSearchResDTO;
 
 public class HomeRecordConverter {
