+++ src/main/java/com/umc/finly/domain/home/dto/response/HomeWeeklyMoodResDTO.java
@@ -1,4 +1,4 @@
-package com.umc.finly.domain.home.dto.res;
+package com.umc.finly.domain.home.dto.response;
 
 import com.umc.finly.domain.record.enums.EmotionCode;
 import lombok.Builder;
