+++ src/main/java/com/umc/finly/domain/home/service/HomeMindServiceImpl.java
@@ -3,8 +3,8 @@ package com.umc.finly.domain.home.service;
 import com.umc.finly.domain.analysis.association.entity.ConvictionScoreResult;
 import com.umc.finly.domain.analysis.association.entity.FearIndexResult;
 import com.umc.finly.domain.home.converter.HomeMindConverter;
-import com.umc.finly.domain.home.dto.res.HomeMindDetailResDTO;
-import com.umc.finly.domain.home.dto.res.HomeMindResDTO;
+import com.umc.finly.domain.home.dto.response.HomeMindDetailResDTO;
+import com.umc.finly.domain.home.dto.response.HomeMindResDTO;
 import com.umc.finly.domain.home.exception.HomeErrorCode;
 import com.umc.finly.domain.home.repository.HomeMindRepository;
 import com.umc.finly.domain.member.entity.Member;
