+++ src/main/java/com/umc/finly/domain/home/service/HomeWeeklyMoodService.java
@@ -1,6 +1,6 @@
 package com.umc.finly.domain.home.service;
 
-import com.umc.finly.domain.home.dto.res.HomeWeeklyMoodResDTO;
+import com.umc.finly.domain.home.dto.response.HomeWeeklyMoodResDTO;
 
 public interface HomeWeeklyMoodService {
