+++ src/main/java/com/umc/finly/domain/market/dto/response/MarketIndexResDTO.java
@@ -1,4 +1,4 @@
-package com.umc.finly.domain.market.dto.res;
+package com.umc.finly.domain.market.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
