+++ src/main/java/com/umc/finly/domain/market/dto/response/MarketInsightResDTO.java
@@ -1,4 +1,4 @@
-package com.umc.finly.domain.market.dto.res;
+package com.umc.finly.domain.market.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
