+++ src/main/java/com/umc/finly/domain/market/infra/MarketIndexProvider.java
@@ -1,6 +1,6 @@
 package com.umc.finly.domain.market.infra;
 
-import com.umc.finly.domain.market.dto.res.MarketIndicesDTO;
+import com.umc.finly.domain.market.dto.response.MarketIndicesDTO;
 
 public interface MarketIndexProvider {
     MarketIndicesDTO getMarketIndices();
